Adding internal identity test file and make target
